Fix XSS (no working exploit found)
This is adds some checks in form of .replace function calls on some user input. I was not able to exploit the issue in a security-relevant way, i only was able to render the site completly unresponsive.

Tried exploits:
" classs="first-name" type="text" placeholder="Vorname"><script>console.log('Hello World')</script><input value="
If you click on the "Neue Person" button, the script tag gets in the HTML, the code inside of it gets never executed, so you only can only break the site, without the possibility to execute any code.

<script>console.log('Hello World')</script>
If you print the document, the field is empty, but no code gets executed

;
If you export it to a CSV-file you can add an arbitrary number of fields.

\n
Did nothing, was included in the CSV as \n and did not cause a new line.